Make_Slerp

+++ source/bsys/file/file.cpp
@@ -79,7 +79,7 @@ namespace NBsys{namespace NFile
 	*/
 	sharedptr< ThreadTemplate< File_Thread > >& GetSystemInstance(s32 a_device_index)
 	{
-		ASSERT((0 <= a_device_index)&&(a_device_index < s_thread_list->size()));
+		ASSERT((0 <= a_device_index)&&(a_device_index < static_cast<s32>(s_thread_list->size())));
 
 		return s_thread_list->at(a_device_index);
 	}
